+++ lib/slimmer.rb
@@ -43,7 +43,6 @@ module Slimmer
     autoload :HeaderContextInserter, 'slimmer/processors/header_context_inserter'
     autoload :InsideHeaderInserter, 'slimmer/processors/inside_header_inserter'
     autoload :NavigationMover, 'slimmer/processors/navigation_mover'
-    autoload :ReportAProblemInserter, 'slimmer/processors/report_a_problem_inserter'
     autoload :SearchIndexSetter, 'slimmer/processors/search_index_setter'
     autoload :SearchPathSetter, 'slimmer/processors/search_path_setter'
     autoload :SearchParameterInserter, 'slimmer/processors/search_parameter_inserter'