+++ test/test_helper.rb
@@ -88,7 +88,6 @@ class SlimmerIntegrationTest < MiniTest::Test
                     end
 
     use_template(template_name)
-    use_template('report_a_problem.raw')
 
     fetch_page
   end